refactor: use default setup to make use of expontiel backoff and add http logging
