+++ internal/controller/account/servicebinding/zz_setup.go
@@ -1,0 +1,33 @@
+package servicebinding
+
+import (
+	"github.com/crossplane/crossplane-runtime/pkg/controller"
+	"github.com/crossplane/crossplane-runtime/pkg/reconciler/managed"
+	"github.com/crossplane/crossplane-runtime/pkg/resource"
+	"github.com/sap/crossplane-provider-btp/apis/account/v1alpha1"
+	"github.com/sap/crossplane-provider-btp/btp"
+	sbClient "github.com/sap/crossplane-provider-btp/internal/clients/account/servicebinding"
+	tfClient "github.com/sap/crossplane-provider-btp/internal/clients/tfclient"
+	"github.com/sap/crossplane-provider-btp/internal/controller/providerconfig"
+	"github.com/sap/crossplane-provider-btp/internal/tracking"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	ctrl "sigs.k8s.io/controller-runtime"
+
+	providerv1alpha1 "github.com/sap/crossplane-provider-btp/apis/v1alpha1"
+)
+
+// Setup adds a controller that reconciles ServiceBinding managed resources.
+func Setup(mgr ctrl.Manager, o controller.Options) error {
+	return providerconfig.DefaultSetup(mgr, o, &v1alpha1.ServiceBinding{}, v1alpha1.ServiceBindingGroupKind, v1alpha1.ServiceBindingGroupVersionKind, func(kube client.Client, usage resource.Tracker, resourcetracker tracking.ReferenceResolverTracker, newServiceFn func(cisSecretData []byte, serviceAccountSecretData []byte) (*btp.Client, error)) managed.ExternalConnecter {
+		return &connector{
+			kube:  mgr.GetClient(),
+			usage: resource.NewProviderConfigUsageTracker(mgr.GetClient(), &providerv1alpha1.ProviderConfigUsage{}),
+
+			//TODO:  put into dedicated function
+			clientConnector: func(kube client.Client) tfClient.TfProxyConnectorI[*v1alpha1.ServiceBinding] {
+				return sbClient.NewServiceBindingConnector(saveCallback, kube)
+			}(mgr.GetClient()),
+		}
+	})
+}
